Adding Game Project
This is one of the first games I created when using CodeHs. The Idea is that your trying to avoid objects that fly toward you with the mouse. Very simple and maybe ill come back to this idea. Also all of the variables had been rearranged to all be on the top because I was planning on adding it into another program.